chore: fix typescript errors
